Refactor utils/utils, fixes #11923
