+++ api/client/ps.go
@@ -15,8 +15,8 @@ import (
 	flag "github.com/docker/docker/pkg/mflag"
 	"github.com/docker/docker/pkg/parsers/filters"
 	"github.com/docker/docker/pkg/stringid"
+	"github.com/docker/docker/pkg/stringutils"
 	"github.com/docker/docker/pkg/units"
-	"github.com/docker/docker/utils"
 )
 
 // CmdPs outputs a list of Docker containers.
@@ -135,7 +135,7 @@ func (cli *DockerCli) CmdPs(args ...string) error {
 		)
 
 		if !*noTrunc {
-			command = utils.Trunc(command, 20)
+			command = stringutils.Truncate(command, 20)
 
 			// only display the default name for the container with notrunc is passed
 			for _, name := range names {
